requirements.txt
flask==2.3.3
flask-cors==4.0.0
torch==2.0.1
torchvision==0.15.2
pillow==10.0.1
numpy==1.24.3
requests==2.31.0
werkzeug==2.3.7
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.1
transformers==4.41.2
